requirements.txt
selenium==4.10.0
python-dotenv==0.21.0
requests==2.31.0
selenium-stealth==1.0.6
webdriver-manager==3.8.6
